+++ src/login.ts
@@ -1,16 +1,10 @@
-import _ from "lodash";
 import Bluebird from "bluebird";
-import inquirer, { QuestionCollection, Question } from "inquirer";
+import inquirer, { QuestionCollection } from "inquirer";
 import zlib from "zlib";
 import { STS, STSClientConfig } from "@aws-sdk/client-sts";
 import { load } from "cheerio";
 import { v4 } from "uuid";
-import puppeteer, {
-  Page,
-  ElementHandle,
-  Browser,
-  HTTPRequest,
-} from "puppeteer";
+import puppeteer, { Browser, HTTPRequest } from "puppeteer";
 import querystring from "querystring";
 import _debug from "debug";
 import { CLIError } from "./CLIError";
@@ -18,8 +12,10 @@ import { awsConfig, ProfileConfig } from "./awsConfig";
 import { HttpsProxyAgent } from "https-proxy-agent";
 import { paths } from "./paths";
 import mkdirp from "mkdirp";
+import fs from "fs/promises";
 import { Agent } from "https";
 import { NodeHttpHandler } from "@smithy/node-http-handler";
+import { states } from "./loginStates";
 
 const debug = _debug("az2aws");
 
@@ -34,410 +30,16 @@ const AWS_SAML_ENDPOINT = "https://signin.aws.amazon.com/saml";
 const AWS_CN_SAML_ENDPOINT = "https://signin.amazonaws.cn/saml";
 const AWS_GOV_SAML_ENDPOINT = "https://signin.amazonaws-us-gov.com/saml";
 
+const getProxyUrl = (): string | undefined =>
+  process.env.https_proxy ||
+  process.env.HTTPS_PROXY ||
+  process.env.http_proxy ||
+  process.env.HTTP_PROXY;
+
 interface Role {
   roleArn: string;
   principalArn: string;
 }
-
-/**
- * To proxy the input/output of the Azure login page, it's easiest to run a loop that
- * monitors the state of the page and then perform the corresponding CLI behavior.
- * The states have a name that is used for the debug messages, a selector that is used
- * with puppeteer's page.$(selector) to determine if the state is active, and a handler
- * that is called if the state is active.
- */
-const states = [
-  {
-    name: "username input",
-    selector: `input[name="loginfmt"]:not(.moveOffScreen)`,
-    async handler(
-      page: Page,
-      _selected: ElementHandle,
-      noPrompt: boolean,
-      defaultUsername: string
-    ): Promise<void> {
-      const error = await page.$(".alert-error");
-      if (error) {
-        debug("Found error message. Displaying");
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-        const errorMessage = await page.evaluate(
-          // eslint-disable-next-line
-          (err) => err?.textContent ?? "",
-          error
-        );
-        console.log(errorMessage);
-      }
-
-      let username;
-
-      if (noPrompt && defaultUsername) {
-        debug("Not prompting user for username");
-        username = defaultUsername;
-      } else {
-        debug("Prompting user for username");
-        ({ username } = await inquirer.prompt([
-          {
-            name: "username",
-            message: "Username:",
-            default: defaultUsername,
-          } as Question,
-        ]));
-      }
-
-      debug("Waiting for username input to be visible");
-      await page.waitForSelector(`input[name="loginfmt"]`, {
-        visible: true,
-        timeout: 60000,
-      });
-
-      debug("Focusing on username input");
-      await page.focus(`input[name="loginfmt"]`);
-
-      debug("Clearing input");
-      for (let i = 0; i < 100; i++) {
-        await page.keyboard.press("Backspace");
-      }
-
-      debug("Typing username");
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-      await page.keyboard.type(username);
-
-      await Bluebird.delay(500);
-
-      debug("Waiting for submit button to be visible");
-      await page.waitForSelector(`input[type=submit]`, {
-        visible: true,
-        timeout: 60000,
-      });
-
-      debug("Submitting form");
-      await page.click("input[type=submit]");
-
-      await Bluebird.delay(500);
-
-      debug("Waiting for submission to finish");
-      await Promise.race([
-        page.waitForSelector(
-          `input[name=loginfmt].has-error,input[name=loginfmt].moveOffScreen`,
-          { timeout: 60000 }
-        ),
-        (async (): Promise<void> => {
-          await Bluebird.delay(1000);
-          await page.waitForSelector(`input[name=loginfmt]`, {
-            hidden: true,
-            timeout: 60000,
-          });
-        })(),
-      ]);
-    },
-  },
-  {
-    name: "account selection",
-    selector: `#aadTile > div > div.table-cell.tile-img > img`,
-    async handler(page: Page): Promise<void> {
-      debug("Multiple accounts associated with username.");
-      const aadTile = await page.$("#aadTileTitle");
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const aadTileMessage: string = await page.evaluate(
-        // eslint-disable-next-line
-        (a) => a?.textContent ?? "",
-        aadTile
-      );
-
-      const msaTile = await page.$("#msaTileTitle");
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const msaTileMessage: string = await page.evaluate(
-        // eslint-disable-next-line
-        (m) => m?.textContent ?? "",
-        msaTile
-      );
-
-      const accounts = [
-        aadTile ? { message: aadTileMessage, selector: "#aadTileTitle" } : null,
-        msaTile ? { message: msaTileMessage, selector: "#msaTileTitle" } : null,
-      ].filter((a): a is { message: string; selector: string } => a !== null);
-
-      let account;
-      if (accounts.length === 0) {
-        throw new CLIError("No accounts found on account selection screen.");
-      } else if (accounts.length === 1) {
-        account = accounts[0];
-      } else {
-        debug("Asking user to choose account");
-        console.log(
-          "It looks like this Username is used with more than one account from Microsoft. Which one do you want to use?"
-        );
-        const answers = await inquirer.prompt([
-          {
-            name: "account",
-            message: "Account:",
-            type: "list",
-            choices: _.map(accounts, "message"),
-            default: aadTileMessage,
-          } as Question,
-        ]);
-
-        account = _.find(accounts, ["message", answers.account]);
-      }
-
-      if (!account) {
-        throw new Error("Unable to find account");
-      }
-
-      debug(`Proceeding with account ${account.selector}`);
-      await page.click(account.selector);
-      await Bluebird.delay(500);
-    },
-  },
-  {
-    name: "passwordless",
-    selector: `input[value='Send notification']`,
-    async handler(page: Page) {
-      debug("Sending notification");
-      // eslint-disable-next-line
-      await page.click("input[value='Send notification']");
-      debug("Waiting for auth code");
-      // eslint-disable-next-line
-      await page.waitForSelector(`#idRemoteNGC_DisplaySign`, {
-        visible: true,
-        timeout: 60000,
-      });
-      debug("Printing the message displayed");
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const messageElement = await page.$(
-        "#idDiv_RemoteNGC_PollingDescription"
-      );
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const codeElement = await page.$("#idRemoteNGC_DisplaySign");
-      // eslint-disable-next-line
-      const message = await page.evaluate(
-        // eslint-disable-next-line
-        (el) => el?.textContent ?? "",
-        messageElement
-      );
-      console.log(message);
-      debug("Printing the auth code");
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const authCode = await page.evaluate(
-        // eslint-disable-next-line
-        (el) => el?.textContent ?? "",
-        codeElement
-      );
-      console.log(authCode);
-      debug("Waiting for response");
-      await page.waitForSelector(`#idRemoteNGC_DisplaySign`, {
-        hidden: true,
-        timeout: 60000,
-      });
-    },
-  },
-  {
-    name: "password input",
-    selector: `input[name="Password"]:not(.moveOffScreen),input[name="passwd"]:not(.moveOffScreen)`,
-    async handler(
-      page: Page,
-      _selected: ElementHandle,
-      noPrompt: boolean,
-      _defaultUsername: string,
-      defaultPassword: string
-    ): Promise<void> {
-      const error = await page.$(".alert-error");
-      if (error) {
-        debug("Found error message. Displaying");
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-        const errorMessage = await page.evaluate(
-          // eslint-disable-next-line
-          (err) => err?.textContent ?? "",
-          error
-        );
-        console.log(errorMessage);
-        defaultPassword = ""; // Password error. Unset the default and allow user to enter it.
-      }
-
-      let password;
-
-      if (noPrompt && defaultPassword) {
-        debug("Not prompting user for password");
-        password = defaultPassword;
-      } else {
-        debug("Prompting user for password");
-        ({ password } = await inquirer.prompt([
-          {
-            name: "password",
-            message: "Password:",
-            type: "password",
-          } as Question,
-        ]));
-      }
-
-      debug("Focusing on password input");
-      await page.focus(`input[name="Password"],input[name="passwd"]`);
-
-      debug("Typing password");
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-      await page.keyboard.type(password);
-
-      debug("Submitting form");
-      await page.click("span[class=submit],input[type=submit]");
-
-      debug("Waiting for a delay");
-      await Bluebird.delay(500);
-    },
-  },
-  {
-    name: "TFA instructions",
-    selector: `#idDiv_SAOTCAS_Description`,
-    async handler(page: Page, selected: ElementHandle): Promise<void> {
-      const descriptionMessage = await page.evaluate(
-        // eslint-disable-next-line
-        (description) => description?.textContent ?? "",
-        selected
-      );
-      console.log(descriptionMessage);
-
-      try {
-        debug("Waiting for authentication code to be displayed");
-        await page.waitForSelector("#idRichContext_DisplaySign", {
-          visible: true,
-          timeout: 5000,
-        });
-        debug("Checking if authentication code is displayed");
-        const authenticationCodeElement = await page.$(
-          "#idRichContext_DisplaySign"
-        );
-        debug("Reading the authentication code");
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-        const authenticationCode = await page.evaluate(
-          // eslint-disable-next-line
-          (d) => d?.textContent ?? "",
-          authenticationCodeElement
-        );
-        debug("Printing the authentication code to console");
-        console.log(authenticationCode);
-      } catch {
-        debug("No authentication code found on page");
-      }
-
-      debug("Waiting for response");
-      await page.waitForSelector(`#idDiv_SAOTCAS_Description`, {
-        hidden: true,
-        timeout: 60000,
-      });
-    },
-  },
-  {
-    name: "TFA failed",
-    selector: `#idDiv_SAASDS_Description,#idDiv_SAASTO_Description`,
-    async handler(page: Page, selected: ElementHandle): Promise<void> {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const descriptionMessage = await page.evaluate(
-        // eslint-disable-next-line
-        (description) => description?.textContent ?? "",
-        selected
-      );
-      throw new CLIError(descriptionMessage);
-    },
-  },
-  {
-    name: "TFA code input",
-    selector: "input[name=otc]:not(.moveOffScreen)",
-    async handler(page: Page): Promise<void> {
-      const error = await page.$(".alert-error");
-      if (error) {
-        debug("Found error message. Displaying");
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-        const errorMessage = await page.evaluate(
-          // eslint-disable-next-line
-          (err) => err?.textContent ?? "",
-          error
-        );
-        console.log(errorMessage);
-      } else {
-        const description = await page.$("#idDiv_SAOTCC_Description");
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-        const descriptionMessage = await page.evaluate(
-          // eslint-disable-next-line
-          (d) => d?.textContent ?? "",
-          description
-        );
-        console.log(descriptionMessage);
-      }
-
-      const { verificationCode } = await inquirer.prompt([
-        {
-          name: "verificationCode",
-          message: "Verification Code:",
-        } as Question,
-      ]);
-
-      debug("Focusing on verification code input");
-      await page.focus(`input[name="otc"]`);
-
-      debug("Clearing input");
-      for (let i = 0; i < 100; i++) {
-        await page.keyboard.press("Backspace");
-      }
-
-      debug("Typing verification code");
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-      await page.keyboard.type(verificationCode);
-
-      debug("Submitting form");
-      await page.click("input[type=submit]");
-
-      debug("Waiting for submission to finish");
-      await Promise.race([
-        page.waitForSelector(
-          `input[name=otc].has-error,input[name=otc].moveOffScreen`,
-          { timeout: 60000 }
-        ),
-        (async (): Promise<void> => {
-          await Bluebird.delay(1000);
-          await page.waitForSelector(`input[name=otc]`, {
-            hidden: true,
-            timeout: 60000,
-          });
-        })(),
-      ]);
-    },
-  },
-  {
-    name: "Remember me",
-    selector: `#KmsiDescription`,
-    async handler(
-      page: Page,
-      _selected: ElementHandle,
-      _noPrompt: boolean,
-      _defaultUsername: string,
-      _defaultPassword: string | undefined,
-      rememberMe: boolean
-    ): Promise<void> {
-      if (rememberMe) {
-        debug("Clicking remember me button");
-        await page.click("#idSIButton9");
-      } else {
-        debug("Clicking don't remember button");
-        await page.click("#idBtn_Back");
-      }
-
-      debug("Waiting for a delay");
-      await Bluebird.delay(500);
-    },
-  },
-  {
-    name: "Service exception",
-    selector: "#service_exception_message",
-    async handler(page: Page, selected: ElementHandle): Promise<void> {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const descriptionMessage = await page.evaluate(
-        // eslint-disable-next-line
-        (description) => description?.textContent ?? "",
-        selected
-      );
-      throw new CLIError(descriptionMessage);
-    },
-  },
-];
 
 export const login = {
   async loginAsync(
@@ -466,6 +68,16 @@ export const login = {
     }
 
     const profile = await this._loadProfileAsync(profileName);
+    console.log(
+      `Using AWS region ${profile.region || "(from AWS SDK defaults)"}`
+    );
+    if (profile.region && profile.region.startsWith("us-gov")) {
+      console.warn(
+        "GovCloud region detected in profile. Note: Other AWS CLI operations " +
+          "will use your AWS CLI default region. If needed, set it to match " +
+          "this GovCloud region (us-gov-west-1 or us-gov-east-1)."
+      );
+    }
     let assertionConsumerServiceURL = AWS_SAML_ENDPOINT;
     if (profile.region && profile.region.startsWith("us-gov")) {
       assertionConsumerServiceURL = AWS_GOV_SAML_ENDPOINT;
@@ -717,8 +329,9 @@ export const login = {
         }
       }
 
-      if (process.env.https_proxy) {
-        args.push(`--proxy-server=${process.env.https_proxy}`);
+      const proxyUrl = getProxyUrl();
+      if (proxyUrl) {
+        args.push(`--proxy-server=${proxyUrl}`);
       }
 
       const ignoreDefaultArgs = noDisableExtensions
@@ -744,7 +357,28 @@ export const login = {
         launchParams.executablePath = paths.chromeBin;
       }
 
-      browser = await puppeteer.launch(launchParams);
+      try {
+        browser = await puppeteer.launch(launchParams);
+      } catch (e) {
+        if (
+          e instanceof Error &&
+          e.name === "TargetCloseError" &&
+          rememberMe &&
+          !paths.userDataDir
+        ) {
+          debug(
+            `Browser launch failed with TargetCloseError. Resetting profile at ${paths.chromium}`
+          );
+          console.warn(
+            "Browser profile appears incompatible. Resetting profile data and retrying..."
+          );
+          await fs.rm(paths.chromium, { recursive: true, force: true });
+          await mkdirp(paths.chromium);
+          browser = await puppeteer.launch(launchParams);
+        } else {
+          throw e;
+        }
+      }
 
       // Wait for a bit as sometimes the browser isn't ready.
       await Bluebird.delay(200);
@@ -956,7 +590,13 @@ export const login = {
     durationHours: number;
   }> {
     let role;
-    let durationHours = parseInt(defaultDurationHours, 10) || 1;
+    let durationHours = 1;
+    if (defaultDurationHours) {
+      const parsedDuration = parseInt(defaultDurationHours, 10);
+      if (!Number.isNaN(parsedDuration) && parsedDuration > 0) {
+        durationHours = parsedDuration;
+      }
+    }
     const questions: QuestionCollection[] = [];
     if (roles.length === 0) {
       throw new CLIError("No roles found in SAML response.");
@@ -964,11 +604,19 @@ export const login = {
       debug("Choosing the only role in response");
       role = roles[0];
     } else {
-      if (noPrompt && defaultRoleArn) {
-        role = _.find(roles, ["roleArn", defaultRoleArn]);
-      }
+      if (noPrompt) {
+        if (!defaultRoleArn) {
+          throw new CLIError(
+            "--no-prompt requires azure_default_role_arn when multiple roles are available."
+          );
+        }
 
-      if (role) {
+        role = roles.find((r) => r.roleArn === defaultRoleArn);
+        if (!role) {
+          throw new CLIError(
+            `Default role ARN '${defaultRoleArn}' was not found in the SAML response.`
+          );
+        }
         debug("Valid role found. No need to ask.");
       } else {
         debug("Asking user to choose role");
@@ -976,14 +624,18 @@ export const login = {
           name: "role",
           message: "Role:",
           type: "list",
-          choices: _.sortBy(_.map(roles, "roleArn")),
+          choices: roles.map((r) => r.roleArn).sort(),
           default: defaultRoleArn,
         });
       }
     }
 
-    if (noPrompt && defaultDurationHours) {
-      debug("Default durationHours found. No need to ask.");
+    if (noPrompt) {
+      if (!defaultDurationHours) {
+        debug("No default durationHours set. Using 1 hour.");
+      } else {
+        debug("Default durationHours found. No need to ask.");
+      }
     } else {
       questions.push({
         name: "durationHours",
@@ -993,7 +645,7 @@ export const login = {
         validate: (input): boolean | string => {
           input = Number(input);
           if (input > 0 && input <= 12) return true;
-          return "Duration hours must be between 0 and 12";
+          return "Duration hours must be between 1 and 12";
         },
       });
     }
@@ -1002,7 +654,7 @@ export const login = {
     // user is logged in and using multiple profiles --all-profiles and --no-prompt
     if (questions.length > 0) {
       const answers = await inquirer.prompt(questions);
-      if (!role) role = _.find(roles, ["roleArn", answers.role]);
+      if (!role) role = roles.find((r) => r.roleArn === answers.role);
       if (answers.durationHours) {
         durationHours = parseInt(answers.durationHours as string, 10);
       }
@@ -1045,15 +697,13 @@ export const login = {
       );
     }
 
-    if (process.env.https_proxy) {
+    const proxyUrl = getProxyUrl();
+    if (proxyUrl) {
       const proxyOptions = awsNoVerifySsl ? { rejectUnauthorized: false } : {};
       stsOptions = {
         ...stsOptions,
         requestHandler: new NodeHttpHandler({
-          httpsAgent: new HttpsProxyAgent(
-            process.env.https_proxy,
-            proxyOptions
-          ),
+          httpsAgent: new HttpsProxyAgent(proxyUrl, proxyOptions),
         }),
       };
     } else if (awsNoVerifySsl) {